server/requirements.txt
asgiref==3.8.1
Django==5.2.3
sqlparse==0.5.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
PyJWT==2.9.0